# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Liberfly API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "User Login here",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "email@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "test123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbGliZXJmbHkudGVzdC9hcGkvbG9naW4iLCJpYXQiOjE2ODAxMjk2ODUsImV4cCI6MTY4MDEzMzI4NSwibmJmIjoxNjgwMTI5Njg1LCJqdGkiOiJjUlZEdzE5U2hRaGRBdXFUIiwic3ViIjoiNSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.GIb7U9y-SbSZK5pebMgVAGulIWo_7U6sUppshrAh6rY"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "email@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "test123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbGliZXJmbHkudGVzdC9hcGkvbG9naW4iLCJpYXQiOjE2ODAxMjk2ODUsImV4cCI6MTY4MDEzMzI4NSwibmJmIjoxNjgwMTI5Njg1LCJqdGkiOiJjUlZEdzE5U2hRaGRBdXFUIiwic3ViIjoiNSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.GIb7U9y-SbSZK5pebMgVAGulIWo_7U6sUppshrAh6rY"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "User authenticated can get a list of all users",
                "description": "User authenticated can get a list of all users",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "users": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "integer",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "email@test.com"
                                                },
                                                "verified_at": {
                                                    "type": "null",
                                                    "example": null
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-03-29T22:29:23.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-03-29T22:29:23.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "User authenticated can get a determinated user",
                "description": "User authenticated can get a determinated user",
                "operationId": "find",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "users": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "name": {
                                                    "type": "integer",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "johndoe@test.com"
                                                },
                                                "verified_at": {
                                                    "type": "null",
                                                    "example": null
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-03-29T22:29:23.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-03-29T22:29:23.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}